pyproject.toml
[tool.poetry]
name = "kiuc-college"
version = "0.1.0"
description = ""
authors = ["Nursultan301 <abykeevnursultan14@gmail.com>"]
readme = "README.md"
packages = [{include = "kiuc_college"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
transliterate = "^1.10.2"
django-imagekit = "^4.1.0"
django-ckeditor = "^6.5.1"
pillow = "^9.5.0"
django-environs = "^0.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
